# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Storytail API",
        "description": "API documentation for Storytail",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/books/{bookId}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Comentários de um livro",
                "description": "Retorna todos os comentários aprovados de um livro",
                "operationId": "5548b168f53f4d8c03c1b9812e0d6ffe",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID do livro",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de comentários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "book_title": {
                                            "type": "string",
                                            "example": "O Grande Livro"
                                        },
                                        "comments": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "comment_text": {
                                                        "type": "string",
                                                        "example": "Excelente livro!"
                                                    },
                                                    "rating": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "moderation_status": {
                                                        "type": "string",
                                                        "example": "approved"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Criar comentário",
                "description": "Adiciona um novo comentário ao livro",
                "operationId": "44c0983143366f14c0cd0cb6ae715eec",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID do livro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "comment_text": {
                                        "type": "string",
                                        "maxLength": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comentário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Comentário submetido com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação ou progresso insuficiente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}